refactor: 역조회, 노선조회함수를 Output클래스로 이동
